GETP-203 refactor: ClientService를 use case에 따라 분리
